+++ project/package.json
@@ -21,7 +21,9 @@
     "@angular/platform-browser": "2.0.0-rc.4",
     "@angular/platform-browser-dynamic": "2.0.0-rc.4",
     "@angular/router": "3.0.0-beta.2",
+    "angularfire2": "^2.0.0-beta.6",
     "es6-shim": "0.35.1",
+    "firebase": "^3.6.4",
     "reflect-metadata": "0.1.3",
     "rxjs": "5.0.0-beta.6",
     "systemjs": "0.19.26",
