+++ project/src/typings.d.ts
@@ -1,6 +1,7 @@
 // Typings reference file, see links for more information
 // https://github.com/typings/typings
 // https://www.typescriptlang.org/docs/handbook/writing-declaration-files.html
-
+/// <reference path="../node_modules/firebase/firebase.d.ts" />
 /// <reference path="../typings/index.d.ts" />
 declare var module: { id: string };
+declare var System: any;
